requirements.txt
streamlit>=1.36
pandas>=2.0
numpy>=1.24
python-dateutil>=2.9
sqlalchemy>=2.0
psycopg[binary]>=3.1
openpyxl>=3.1
xlsxwriter>=3.2
